Split transfer generator factory to many factories
