+++ src/TransferGenerator/Definition/DefinitionFactory.php
@@ -4,243 +4,45 @@ declare(strict_types=1);
 
 namespace Picamator\TransferObject\TransferGenerator\Definition;
 
-use ArrayObject;
-use Picamator\TransferObject\Shared\SharedFactoryTrait;
-use Picamator\TransferObject\TransferGenerator\Config\ConfigFactoryTrait;
-use Picamator\TransferObject\TransferGenerator\Definition\Filesystem\DefinitionFinder;
+use Picamator\TransferObject\Shared\CachedFactoryTrait;
 use Picamator\TransferObject\TransferGenerator\Definition\Filesystem\DefinitionFinderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\EmbeddedTypeBuilder;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\EmbeddedTypeBuilderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\NamespaceBuilder;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\NamespaceBuilderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\ContentBuilder;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\ContentBuilderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\DefinitionParser;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\DefinitionParserInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\CollectionTypePropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\DateTimeTypePropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\EnumTypePropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\NullablePropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\NumberTypePropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\PropertyExpanderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\ProtectedPropertyExpander;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\TypePropertyExpander;
+use Picamator\TransferObject\TransferGenerator\Definition\Parser\ParserFactory;
 use Picamator\TransferObject\TransferGenerator\Definition\Reader\DefinitionReader;
 use Picamator\TransferObject\TransferGenerator\Definition\Reader\DefinitionReaderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Content\ClassNameContentValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Content\ContentValidatorInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Content\EmptyPropertiesContentValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Content\PropertiesContentValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\DefinitionValidator;
 use Picamator\TransferObject\TransferGenerator\Definition\Validator\DefinitionValidatorInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\BuildInTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\CollectionTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\DateTimeTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\EnumTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\NamePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\NumberTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\PropertyValidatorInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\RequiredTypePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\ReservedNamePropertyValidator;
-use Picamator\TransferObject\TransferGenerator\Definition\Validator\Property\TransferTypePropertyValidator;
+use Picamator\TransferObject\TransferGenerator\Definition\Validator\ValidatorFactory;
 
 class DefinitionFactory
 {
-    use ConfigFactoryTrait;
-    use SharedFactoryTrait;
-
-    private DefinitionReaderInterface $definitionReader;
+    use CachedFactoryTrait;
 
     public function createDefinitionReader(): DefinitionReaderInterface
     {
-        return $this->definitionReader ??= new DefinitionReader(
-            $this->createDefinitionFinder(),
-            $this->createDefinitionParser(),
-            $this->createContentValidator(),
+        /** @phpstan-ignore return.type */
+        return $this->getCached(
+            key: 'definition-reader',
+            factory: fn (): DefinitionReaderInterface =>
+                new DefinitionReader(
+                    $this->createDefinitionFinder(),
+                    $this->createDefinitionParser(),
+                    $this->createDefinitionValidator(),
+                ),
         );
     }
 
-    protected function createContentValidator(): DefinitionValidatorInterface
+    protected function createDefinitionValidator(): DefinitionValidatorInterface
     {
-        return new DefinitionValidator(
-            $this->createContentValidators(),
-        );
-    }
-
-    /**
-     * @return \ArrayObject<int,ContentValidatorInterface>
-     */
-    protected function createContentValidators(): ArrayObject
-    {
-        return new ArrayObject([
-            $this->createClassNameContentValidator(),
-            $this->createEmptyPropertiesContentValidator(),
-            $this->createPropertiesContentValidator(),
-        ]);
-    }
-
-    protected function createPropertiesContentValidator(): ContentValidatorInterface
-    {
-        return new PropertiesContentValidator($this->createPropertyValidators());
-    }
-
-    /**
-     * @return \ArrayObject<int,PropertyValidatorInterface>
-     */
-    protected function createPropertyValidators(): ArrayObject
-    {
-        return new ArrayObject([
-            $this->createReservedNamePropertyValidator(),
-            $this->createNamePropertyValidator(),
-            $this->createRequiredTypePropertyValidator(),
-            $this->createBuildInTypePropertyValidator(),
-            $this->createTransferTypePropertyValidator(),
-            $this->createCollectionTypePropertyValidator(),
-            $this->createEnumTypePropertyValidator(),
-            $this->createDateTimeTypePropertyValidator(),
-            $this->createNumberTypePropertyValidator(),
-        ]);
-    }
-
-    protected function createNumberTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new NumberTypePropertyValidator();
-    }
-
-    protected function createDateTimeTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new DateTimeTypePropertyValidator();
-    }
-
-    protected function createEnumTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new EnumTypePropertyValidator();
-    }
-
-    protected function createCollectionTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new CollectionTypePropertyValidator(
-            $this->createClassNameValidator(),
-            $this->createNamespaceValidator(),
-        );
-    }
-
-    protected function createTransferTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new TransferTypePropertyValidator(
-            $this->createClassNameValidator(),
-            $this->createNamespaceValidator(),
-        );
-    }
-
-    protected function createBuildInTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new BuildInTypePropertyValidator();
-    }
-
-    protected function createRequiredTypePropertyValidator(): PropertyValidatorInterface
-    {
-        return new RequiredTypePropertyValidator();
-    }
-
-    protected function createNamePropertyValidator(): PropertyValidatorInterface
-    {
-        return new NamePropertyValidator();
-    }
-
-    protected function createReservedNamePropertyValidator(): PropertyValidatorInterface
-    {
-        return new ReservedNamePropertyValidator();
-    }
-
-    protected function createEmptyPropertiesContentValidator(): ContentValidatorInterface
-    {
-        return new EmptyPropertiesContentValidator();
-    }
-
-    protected function createClassNameContentValidator(): ContentValidatorInterface
-    {
-        return new ClassNameContentValidator($this->createClassNameValidator());
-    }
-
-    protected function createDefinitionParser(): DefinitionParserInterface
-    {
-        return new DefinitionParser(
-            $this->getYmlParser(),
-            $this->createContentBuilder(),
-        );
-    }
-
-    protected function createContentBuilder(): ContentBuilderInterface
-    {
-        return new ContentBuilder($this->createPropertyExpander());
-    }
-
-    protected function createPropertyExpander(): PropertyExpanderInterface
-    {
-        $propertyExpander = $this->createNullablePropertyExpander();
-
-        $propertyExpander
-            ->setNextExpander($this->createCollectionTypePropertyExpander())
-            ->setNextExpander($this->createTypePropertyExpander())
-            ->setNextExpander($this->createEnumTypePropertyExpander())
-            ->setNextExpander($this->createProtectedPropertyExpander())
-            ->setNextExpander($this->createDateTimeTypePropertyExpander())
-            ->setNextExpander($this->createNumberTypePropertyExpander());
-
-        return $propertyExpander;
-    }
-
-    protected function createNumberTypePropertyExpander(): PropertyExpanderInterface
-    {
-        return new NumberTypePropertyExpander($this->createEmbeddedTypeBuilder());
-    }
-
-    protected function createEmbeddedTypeBuilder(): EmbeddedTypeBuilderInterface
-    {
-        return new EmbeddedTypeBuilder($this->createNamespaceBuilder());
-    }
-
-    protected function createNamespaceBuilder(): NamespaceBuilderInterface
-    {
-        return new NamespaceBuilder();
-    }
-
-    protected function createDateTimeTypePropertyExpander(): PropertyExpanderInterface
-    {
-        return new DateTimeTypePropertyExpander($this->createEmbeddedTypeBuilder());
-    }
-
-    protected function createProtectedPropertyExpander(): PropertyExpanderInterface
-    {
-        return new ProtectedPropertyExpander();
-    }
-
-    protected function createEnumTypePropertyExpander(): PropertyExpanderInterface
-    {
-        return new EnumTypePropertyExpander($this->createEmbeddedTypeBuilder());
-    }
-
-    protected function createTypePropertyExpander(): PropertyExpanderInterface
-    {
-        return new TypePropertyExpander($this->createEmbeddedTypeBuilder());
-    }
-
-    protected function createCollectionTypePropertyExpander(): PropertyExpanderInterface
-    {
-        return new CollectionTypePropertyExpander($this->createEmbeddedTypeBuilder());
-    }
-
-    protected function createNullablePropertyExpander(): PropertyExpanderInterface
-    {
-        return new NullablePropertyExpander();
+        return new ValidatorFactory()->createDefinitionValidator();
     }
 
     protected function createDefinitionFinder(): DefinitionFinderInterface
     {
-        return new DefinitionFinder(
-            $this->getFinder(),
-            $this->getConfig(),
-        );
+        return new ParserFactory()->createDefinitionFinder();
+    }
+
+    protected function createDefinitionParser(): DefinitionParserInterface
+    {
+        return new ParserFactory()->createDefinitionParser();
     }
 }
